[WIP] Moving the overridden method "render_constraint_element" to the AdvancedHelper
Implementing a separate test suite for the extended Helper
